@josh @ricardo replace the tabindex handling for checkboxes
Using tabindex at div level was shadowing the actual focus on the input. Because of that,
we've just removed the width:0, height:0 on the input, as this hiding mechanism was redundant
considering we already had opacity:0. This change on the box model allows Safari to set the focus
on the checkbox properly.

Please note that the links doesn't get the focus on Safari. To fix that, we'd need to add a tabindex
attribute on the DOM, which is the responsibility, talking about the signup journey, of the
rango-transformer mechanism.